Finish the filter and ordering features
This commit completes the feature list by adding the ability to filter
and sort the companies table.

I first installed Stimulus, a javascript frontend tool that compliments
turbo. This allows us to write javascript controllers for actions on the
front end.

I then took the companies table I wrote on the index page
and refactored it into it's own partial template, as I am planning
to re-render that table from a new endpoint. I also took the opportunity
to wrap the table in a turbo-frame, which leverages the Turbo framework.
That framework intercepts all calls made from within the element and
makes them async calls to the backend. It also provides a target in the
DOM to re-render with the response from that async call.

I then added the ability to sort by ticker symbol and company name by
using the CompaniesHelper (a view helper for methods used when rendering
templates) to generate links for the table headers. These links make a
call to a new 'filter' endpoint using the parameters 'column' and
'direction', which will return a sorted list of companies to re-render
in the company table partial asyncronously via Turbo.

Next I filled out the search sidebar, with a form - adding text input
fields for company name and ticker, and select boxes for each issue area.
When submitted, this form goes to the same filter endpoint, this time
using the submitted form to filter the company table to the companies
that match the criteria. I then wrote an endpoint on my stimulus JS
controller - it is a method that when called, resets a 200ms timer, and
when that timer finishes submits the form. This method gets called on
each text input into the text box and each selection in the select boxes.
The 200 ms timer is to ensure that we don't make a call every time someone
types, and gives us enough time to get the full entry.

The select box was not my first choice, I first used radio switches,
but soon realized that there are three states, not two (True, False, no
preference), so I swtiched to a select box which gave me that ability.

Both the search and filter mechanics were added to the company model as
class methods, both for ensuring correct separation of concerns, and
to allow for more simple testability.

As a note, all filter parameters are saved to the session so that multiple
calls remember each other (i.e. if you sorted the table, and then
filtered it, it would remember the sort). All filter params in the session
are cleared on the initial load of the index page.

Some other small additions include:
- node module for bootstrap icons, so I could use little directional
caret svgs as icons to indicate sort direction.
- Fixed a bug in the company model which counted the issue areas inversely
- Fixed the strong params (and added a new one for filters) from earlier
bad naming for company attributes