Added model testing for Company
Would have loved to do a true TDD and write tests for the controller,
views, and helpers but didn't in the interest of time.

Just tested some core search methods on the company model - as well as
wrote a factory and installed rspec. Found a couple of bugs - lack of
default arguments for the methods led to some bugs. Parameters always
came in fully populated (even with blank strings) so it never came up.

Some testing things I would have done differently with more time:
  - Forgot to start with rspec in the initial generator and didn't
  realize until it was too late
  - Would have leveraged the factory to do more fuzzy testing for the
  model
  - Wouldn't have done count testing for filtering, shortcut for time
  - don't love blank default arguments like I have in the company model